Introduce PoseGraphInterface.
