+++ cartographer/mapping/map_builder.h
@@ -23,6 +23,7 @@
 #include <unordered_map>
 
 #include "cartographer/common/thread_pool.h"
+#include "cartographer/mapping/pose_graph_interface.h"
 #include "cartographer/mapping/proto/map_builder_options.pb.h"
 #include "cartographer/mapping_2d/pose_graph.h"
 #include "cartographer/mapping_3d/pose_graph.h"
@@ -65,7 +66,7 @@ class MapBuilder : public MapBuilderInterface {
 
   int num_trajectory_builders() const override;
 
-  mapping::PoseGraph* pose_graph() override;
+  mapping::PoseGraphInterface* pose_graph() override;
 
  private:
   const proto::MapBuilderOptions options_;
